basic logic of the game working
