+++ src/Game.js
@@ -1,0 +1,5 @@
+class Game {
+    constructor() {
+        this.locations = [France, Espagne, Portugal] // DUMMY DATAS
+    }
+}
